Switched make_key() to deal with three-char keys.
This led to a problem: the command 'baa' is then ambiguous; it could be 'go to
bookmark item aa' or 'go to item baa'. The solution contained in this commit
is to switch the b, g and h commands to uppercase (so Baa [bookmark aa] is now
distinguishable from baa [item baa]).